backend/requirements.txt
asgiref==3.11.0
dataclasses==0.6
dj-database-url==3.1.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.2.31
dnspython==2.8.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pymongo==4.16.0
PyMySQL==1.1.2
sqlparse==0.5.5
whitenoise==6.11.0
